Renaming app step 1
