+++ app/.cxx/Debug/2e13j2y2/arm64-v8a/.ninja_log
@@ -1,5 +1,3 @@
 # ninja log v5
-1	86	653813805	CMakeFiles/demo2.dir/demo2.cpp.o	ad103b95543ac82b
-87	1428	653813807	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/arm64-v8a/libdemo2.so	8d16ac51aa09f878
-1	98	653814081	CMakeFiles/demo2.dir/demo2.cpp.o	ad103b95543ac82b
-98	1394	653814083	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/arm64-v8a/libdemo2.so	8d16ac51aa09f878
+1	117	653816212	CMakeFiles/demo2.dir/demo2.cpp.o	ad103b95543ac82b
+118	227	653816212	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/arm64-v8a/libdemo2.so	8d16ac51aa09f878
