+++ app/.cxx/Debug/2e13j2y2/armeabi-v7a/.ninja_log
@@ -1,5 +1,3 @@
 # ninja log v5
-1	100	653813804	CMakeFiles/demo2.dir/demo2.cpp.o	90bcc2b8a16582bf
-101	1440	653813805	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/armeabi-v7a/libdemo2.so	bbbe693612c108dc
-2	153	653814080	CMakeFiles/demo2.dir/demo2.cpp.o	90bcc2b8a16582bf
-153	1493	653814081	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/armeabi-v7a/libdemo2.so	bbbe693612c108dc
+1	85	653816212	CMakeFiles/demo2.dir/demo2.cpp.o	90bcc2b8a16582bf
+86	198	653816212	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/armeabi-v7a/libdemo2.so	bbbe693612c108dc
