+++ app/.cxx/Debug/2e13j2y2/x86/.ninja_log
@@ -1,5 +1,3 @@
 # ninja log v5
-1	84	653813807	CMakeFiles/demo2.dir/demo2.cpp.o	4d7004ea700df48
-84	1392	653813808	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/x86/libdemo2.so	826b1b9e8a6bf4d2
-1	84	653814083	CMakeFiles/demo2.dir/demo2.cpp.o	4d7004ea700df48
-84	1403	653814084	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/x86/libdemo2.so	826b1b9e8a6bf4d2
+0	79	653816212	CMakeFiles/demo2.dir/demo2.cpp.o	4d7004ea700df48
+80	189	653816212	C:/Users/Kacey/Desktop/Educational-Solar-Sim/app/build/intermediates/cxx/Debug/2e13j2y2/obj/x86/libdemo2.so	826b1b9e8a6bf4d2
