+++ app/src/main/java/com/project/solarsim/MainActivity.java
@@ -1,4 +1,4 @@
-package com.project.demo2;
+package com.project.solarsim;
 
 import android.os.Bundle;
 
@@ -13,7 +13,7 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-import com.project.demo2.databinding.ActivityMainBinding;
+import com.project.solarsim.databinding.ActivityMainBinding;
 
 import android.view.Menu;
 import android.view.MenuItem;
