+++ app/src/test/java/com/project/solarsim/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.project.demo2;
+package com.project.solarsim;
 
 import org.junit.Test;
